requirements.txt
streamlit
praw
google-generativeai
